+++ src/Math.h
@@ -13,8 +13,8 @@ namespace Math {
 
     const double pi = 3.14159265358979;
 
-    int distanceSquared(int ax, int ay, int bx, int by);
-    int distanceSquared(int x, int y);
+    double distanceSquared(double ax, double ay, double bx, double by);
+    double distanceSquared(double x, double y);
 
     double randomRange(double min, double max);
